CLI client application (#17)
This PR adds a feature-complete but extremely simple CLI application to interface with the todos
server. The CLI has very advanced configuration and authentication mechanisms, cacheing 
access tokens in a system configuration directory and refreshing to prevent multiple logins.
Interaction with the API, however, is REST-based only (e.g. no specific commands). 

With the completion of this PR, the server/client is technically ready to be used. 

Version Bump to v1.1

Fixes #2